fix: restrict create-first-team page for logged out users
